Updated geometry with input from Gabriella
Changed some dimensions and positions based on feedback. Recalculated calorimeter rotation based on the fact that the bars below the support are not placed at the far ends of the calorimeter.
Also moved to a G4UnionSolid of bars and support+housing as they are attached to each other (Back bar will lift of the ground when feet are attached in the back).