Changed geometry to implementation of rotation only in calorimeter
The previous geometry did not allow to place the Preshower near the calorimeter without it being rotated along with it, as is the case in the testbeam. Now the air cylinder is gone and the rotation is only applied to the calorimeter.
It is moved such that the centre of rotation aligns with the centre of the calorimeter (same as test beam).
Additionally, the entire geometry has been shifted such that the centre of the calo is now at (0, 0, 0). With no rotation the front face of the calorimeter is therefore centered at (0, 0) in x-y