Added fibre attenuation to simulation
Fibre attenuation based on test beam data is simulated. For each photon emission a probabilty of survival is calculated based on an exponential decay.
Survival of each photon is drawn with this probability which is determined via the distance from creation to SiPM (end of fibre)
For distance to SiPM only the z information of the step along the fibre is used.